requirements.txt
discord.py
python-dotenvpip
Pillow